fix: verifica se usuario tem autorizacao para excluir despesa

+++ app/Http/Controllers/DespesaController.php
@@ -47,6 +47,8 @@ class DespesaController extends Controller
 
     public function destroy(Despesa $despesa)
     {
+        $this->authorize('delete', $despesa);
+        
         $despesa->delete();
     }
 }
